[BONUS] Some bug fixes

+++ libft/ft_nbrlen.c
@@ -6,11 +6,13 @@
 /*   By: aperez-b <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/21 20:16:39 by aperez-b          #+#    #+#             */
-/*   Updated: 2021/04/28 09:33:02 by aperez-b         ###   ########.fr       */
+/*   Updated: 2021/05/11 18:28:05 by aperez-b         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_nbrlen(long n, int base)
+#include <stdio.h>
+
+int	ft_nbrlen(long long unsigned n, int base)
 {
 	int	len;
